+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add shortcuts `new`/`edit` to launch the TUI editor
+- Add output format argument for `todo new` command
 
 ## [v0.2.2] - 2025-07-28